+++ src/pool.rs
@@ -1,15 +1,20 @@
 use std::sync::Mutex;
 
+use crate::migration::Config;
+
 pub(crate) struct Pool {
-    manager: r2d2_sqlite::SqliteConnectionManager,
+    config: Config,
     reserve: Mutex<Vec<rusqlite::Connection>>,
     max_reserve: usize,
+    // extra connection to keep in memory databases alive
+    _anchor: Mutex<rusqlite::Connection>,
 }
 
 impl Pool {
-    pub fn new(manager: r2d2_sqlite::SqliteConnectionManager) -> Self {
+    pub fn new(config: Config) -> Self {
         Self {
-            manager,
+            _anchor: Mutex::new(config.connect().unwrap()),
+            config,
             reserve: Mutex::new(Vec::new()),
             max_reserve: 5,
         }
@@ -17,10 +22,8 @@ impl Pool {
 
     /// Get a new connection from the reserve or make a new one.
     pub fn pop(&self) -> rusqlite::Connection {
-        self.pop_fast().unwrap_or_else(|| {
-            use r2d2::ManageConnection;
-            self.manager.connect().unwrap()
-        })
+        self.pop_fast()
+            .unwrap_or_else(|| self.config.connect().unwrap())
     }
 
     // code optimized to hold lock for shortest time possible